+++ runner/operator/__init__.py
@@ -1,1 +1,5 @@
+from .roslin_operator import roslin_operator
+from .copy_outputs_operator import copy_outputs_operator
+from .tempo_operator import tempo_operator
+from .access_operator import access_operator
 from .operator_factory import OperatorFactory
